Add Roles System

+++ hackpi/Auth.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordBearer
 from pydantic import BaseModel
 from sqlalchemy import Column, Integer, String
 from starlette.responses import JSONResponse
@@ -6,10 +7,15 @@ from starlette.responses import JSONResponse
 from hackpi import Database
 from hackpi.Database import Base
 from hackpi.JWT import JWT
+from hackpi.Methods import Methods
+from hackpi.Roles import StandartRoles
 
+# move it to new file
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 class AbstractUser:
     id = Column(Integer, primary_key=True)
+    role = Column(String, default=StandartRoles.USER)
 
 
 class UserModel(Base, AbstractUser):
@@ -25,7 +31,7 @@ class UserSchema(BaseModel):
 
 
 class Auth:
-    def __init__(self, database: Database, jwt: JWT, model: Base = UserModel, schema: BaseModel = UserSchema):
+    def __init__(self, database: Database, jwt: JWT, model: Base = UserModel, schema: BaseModel = UserSchema, roles: dict[str, list[str]] = {}):
         self.__database = database
         self.__model = model
         self.__schema = schema
@@ -42,39 +48,54 @@ class Auth:
             ))
             session.commit()
 
-            return jwt.create({'username': user.username})
+            return jwt.create({'username': user.username, 'role': StandartRoles.USER})
 
         @self.__router.post('/sign-in')
         def sign_in(user: self.__schema, session=Depends(self.__database.get_session)):
             row = session.query(self.__model).filter(self.__model.username == user.username).one()
 
             if row.password == user.password:
-                return jwt.create({'username': user.username})
+                return jwt.create({'username': user.username, 'role': row.role})
             else:
                 return HTTPException(status_code=401, detail='Invalid credentials.')
 
-        @self.__router.get('/get-users')
-        def get_users(session=Depends(self.__database.get_session)):
-            return session.query(self.__model).all()
+        if roles.get(Methods.GET):
+            @self.__router.get('/get-users')
+            def get_users(session=Depends(self.__database.get_session), token: str = Depends(oauth2_scheme)):
+                if jwt.parse(token).get('role') in roles[Methods.GET]:
+                    return session.query(self.__model).all()
+                else:
+                    return HTTPException(status_code=401, detail='Not permitted')
 
-        @self.__router.get('/get-user-by-id')
-        def get_user_by_id(id: int, session=Depends(self.__database.get_session)):
-            # check if id is exist
-            return session.query(self.__model).filter(self.__model.id == id).one()
+            @self.__router.get('/get-user-by-id')
+            def get_user_by_id(id: int, session=Depends(self.__database.get_session), token: str = Depends(oauth2_scheme)):
+                if jwt.parse(token).get('role') in roles[Methods.GET]:
+                    # check if id is exist
+                    return session.query(self.__model).filter(self.__model.id == id).one()
+                else:
+                    return HTTPException(status_code=401, detail='Not permitted')
 
-        @self.__router.put('/userinfo-update')
-        def userinfo_update(id: int, userinfo: self.__schema, session=Depends(self.__database.get_session)):
-            session.query(self.__model).filter(self.__model.id == id).update(userinfo.__dict__)
-            session.commit()
+        if roles.get(Methods.PUT):
+            @self.__router.put('/userinfo-update')
+            def userinfo_update(id: int, userinfo: self.__schema, session=Depends(self.__database.get_session), token: str = Depends(oauth2_scheme)):
+                if jwt.parse(token).get('role') in roles[Methods.PUT]:
+                    session.query(self.__model).filter(self.__model.id == id).update(userinfo.__dict__)
+                    session.commit()
 
-            return JSONResponse(content={'message': 'The user has been updated'}, status_code=204)
+                    return JSONResponse(content={'message': 'The user has been updated'}, status_code=204)
+                else:
+                    return HTTPException(status_code=401, detail='Not permitted')
 
-        @self.__router.delete('/user-delete')
-        def user_delete(id: int, session=Depends(self.__database.get_session)):
-            session.query(self.__model).filter(self.__model.id == id).delete()
-            session.commit()
+        if roles.get(Methods.DELETE):
+            @self.__router.delete('/user-delete')
+            def user_delete(id: int, session=Depends(self.__database.get_session), token: str = Depends(oauth2_scheme)):
+                if jwt.parse(token).get('role') in roles[Methods.DELETE]:
+                    session.query(self.__model).filter(self.__model.id == id).delete()
+                    session.commit()
 
-            return JSONResponse(content={'message': 'The user has been deleted'}, status_code=200)
+                    return JSONResponse(content={'message': 'The user has been deleted'}, status_code=200)
+                else:
+                    return HTTPException(status_code=401, detail='Not permitted')
 
     def __call__(self):
         return self.__router
